+++ src/app/collections/apiDocumentation/apiDocumentation.controller.js
@@ -17,12 +17,12 @@
 
         function activate () {
             vm.columnSet = [
-                { predicate: 'developer', display: 'Developer' },
-                { predicate: 'product', display: 'Product' },
-                { predicate: 'version', display: 'Version' },
-                { predicate: 'chplProductNumber', display: 'CHPL ID', sortDefault: true, isLink: true },
-                { predicate: 'apiDocumentation', display: 'API Documentation', transformFn: vm.apiTransform },
-                { predicate: 'transparencyAttestationUrl', display: 'Mandatory Disclosures URL', transformFn: vm.disclosuresTransform },
+                { predicate: 'developer', display: 'Developer', sortType: 'single' },
+                { predicate: 'product', display: 'Product', sortType: 'single' },
+                { predicate: 'version', display: 'Version', sortType: 'single' },
+                { predicate: 'chplProductNumber', display: 'CHPL ID', sortType: 'single', sortDefault: true, isLink: true },
+                { predicate: 'apiDocumentation', display: 'API Documentation', sortType: 'single', transformFn: vm.apiTransform },
+                { predicate: 'transparencyAttestationUrl', display: 'Mandatory Disclosures URL', sortType: 'single', transformFn: vm.disclosuresTransform },
             ];
         }