+++ src/app/collections/collections.service.spec.js
@@ -75,5 +75,26 @@
                 expect(results[0].mainSearch).toEqual('DrScribe, Inc.|365EHR|4.0.14|CHP-026059');
             });
         });
+
+        describe('transparencyAttestations translation', function () {
+            var results;
+            beforeEach(function () {
+                results = service.translate('transparencyAttestations', {developers: Mock.developers});
+            });
+
+            it('should filter on transparencyAttestations', function () {
+                expect(results.length).toBe(3);
+            });
+
+            it('should generate a mainSearch', function () {
+                expect(results[0].mainSearch).toEqual('ACL Laboratories');
+            });
+
+            it('should join Attestations', function () {
+                expect(results[0].transparencyAttestation).toEqual('<strong>ICSA Labs</strong>: <span class="text-muted">Not Applicable</span>');
+                expect(results[1].transparencyAttestation).toEqual('<strong>Drummond Group</strong>: <span class="text-success">Supports</span>');
+                expect(results[2].transparencyAttestation).toEqual('<strong>Drummond Group</strong>: <span class="text-success">Supports</span>');
+            });
+        });
     });
 })();