+++ src/app/common/common.service.js
@@ -151,15 +151,17 @@
 
         function getCollection (type) {
             switch (type) {
-            case 'nonconformities':
-                return simpleApiCall('/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,surveillanceCount,openNonconformityCount,closedNonconformityCount');
-            case 'decertifiedProducts':
-            case 'inactiveCertificates':
-                return simpleApiCall('/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,decertificationDate,certificationStatus,numMeaningfulUse');
             case 'apiDocumentation':
                 return simpleApiCall('/certified_products?fields=id,edition,developer,product,version,chplProductNumber,criteriaMet,apiDocumentation,transparencyAttestationUrl');
             case 'bannedDevelopers':
                 return simpleApiCall('/decertifications/developers');
+            case 'decertifiedProducts':
+            case 'inactiveCertificates':
+                return simpleApiCall('/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,decertificationDate,certificationStatus,numMeaningfulUse');
+            case 'nonconformities':
+                return simpleApiCall('/certified_products?fields=id,edition,developer,product,version,chplProductNumber,acb,surveillanceCount,openNonconformityCount,closedNonconformityCount');
+            case 'transparencyAttestations':
+                return simpleApiCall('/developers/');
                 //no default
             }
         }